+++ src/main/java/fr/maxlego08/essentials/commands/commands/fly/CommandFly.java
@@ -27,7 +27,7 @@ public class CommandFly extends VCommand {
 
         Player player = this.argAsPlayer(0, this.player);
 
-        if (this.player == null) {
+        if (player == null) {
             return CommandResultType.SYNTAX_ERROR;
         }