+++ src/main/java/fr/maxlego08/essentials/commands/commands/utils/CommandNightVision.java
@@ -25,7 +25,7 @@ public class CommandNightVision extends VCommand {
 
         Player player = this.argAsPlayer(0, this.player);
 
-        if (this.player == null) {
+        if (player == null) {
             return CommandResultType.SYNTAX_ERROR;
         }