+++ src/main/java/fr/maxlego08/essentials/commands/commands/utils/CommandRepairAll.java
@@ -5,6 +5,7 @@ import fr.maxlego08.essentials.api.commands.CommandResultType;
 import fr.maxlego08.essentials.api.commands.Permission;
 import fr.maxlego08.essentials.api.messages.Message;
 import fr.maxlego08.essentials.zutils.utils.commands.VCommand;
+import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.Damageable;
 import org.bukkit.inventory.meta.ItemMeta;
@@ -14,15 +15,25 @@ public class CommandRepairAll extends VCommand {
         super(plugin);
         this.setPermission(Permission.ESSENTIALS_REPAIR_ALL);
         this.setDescription(Message.DESCRIPTION_REPAIR_ALL);
+        this.addOptionalArg("player");
     }
 
     @Override
     protected CommandResultType perform(EssentialsPlugin plugin) {
 
+        Player player = this.argAsPlayer(0, this.player);
+
+        if (!hasPermission(sender, Permission.ESSENTIALS_REPAIR_ALL_OTHER)) {
+            player = this.player;
+        }
+
+        if (player == null) {
+            return CommandResultType.SYNTAX_ERROR;
+        }
 
         int amount = 0;
 
-        for (ItemStack itemStack : this.player.getInventory().getContents()) {
+        for (ItemStack itemStack : player.getInventory().getContents()) {
             if (itemStack != null) {
                 ItemMeta itemMeta = itemStack.getItemMeta();
                 if (itemMeta instanceof Damageable damageable) {
@@ -36,12 +47,12 @@ public class CommandRepairAll extends VCommand {
         }
 
         if (amount == 0) {
-            message(sender, Message.COMMAND_REPAIR_ALL_ERROR);
+            message(player, Message.COMMAND_REPAIR_ALL_ERROR);
             return CommandResultType.DEFAULT;
         }
 
-        this.player.updateInventory();
-        message(sender, Message.COMMAND_REPAIR_ALL_SUCCESS);
+        player.updateInventory();
+        message(player, Message.COMMAND_REPAIR_ALL_SUCCESS);
         return CommandResultType.SUCCESS;
     }
 }